+++ src/lib/components/Button/ButtonGroup.spec.tsx
@@ -1,8 +1,8 @@
 import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import { describe, expect, it, vi } from 'vitest';
+import { Button } from '.';
 import defaultTheme from '../../theme/default';
-import { Button } from '../Button';
 import { Flowbite } from '../Flowbite';
 
 describe('Components / Button group', () => {
