+++ src/lib/components/Button/index.ts
@@ -1,1 +1,2 @@
 export * from './Button';
+export type { ButtonGroupProps, PositionInButtonGroup } from './ButtonGroup';
