--- src/lib/components/ButtonGroup/index.ts
@@ -1,3 +1,0 @@
-export * from './ButtonGroup';
-export { default } from './ButtonGroup';
-export type { ButtonGroupProps, PositionInButtonGroup } from './ButtonGroup';
